chore: rename SQL file and configure data insertion on startup
Renamed `data-tables.sql` to `data.sql` for consistency with Spring Boot's default initialization file name and moved the file to the root of the `resources` directory to ensure automatic data insertion on application startup.
Updated configuration with `spring.jpa.hibernate.ddl-auto=create` to recreate the schema and ensure that `data.sql` is used to insert initial data into the database after schema creation.